[CLN] Move API to the root folder of gempy
